modified thread sleep wait to allow 0.5 second waits

+++ src/test/java/com/twoitesting/utilityClasses/Waiter.java
@@ -28,9 +28,9 @@ public class Waiter {
         wait.until(ExpectedConditions.visibilityOf(element));
     }
 
-    public static void threadSleep(int seconds) {
+    public static void threadSleep(double seconds) {
 
-        int milliseconds = seconds * 1000;
+        int milliseconds = (int)(seconds * 1000);
 
         try {
             Thread.sleep(milliseconds);
